tests/requirements.txt
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
faker>=19.0.0
httpx>=0.24.0

